# Fix #4457 race condition during FiberContext interruption