--- a/core-tests/jvm/src/test/scala/zio/internal/FiberInterruptSpec.scala
+++ b/core-tests/jvm/src/test/scala/zio/internal/FiberInterruptSpec.scala
@@ -1,12 +1,12 @@
 package zio.internal
 
 import zio.Promise.internal.Pending
+import zio.duration._
 import zio.test.Assertion.equalTo
+import zio.test.TestAspect._
 import zio.test.environment.Live
 import zio.test.{ ZSpec, assertM, suite, testM }
 import zio.{ Promise, ZIOBaseSpec }
-import zio.duration._
-import zio.test.TestAspect._
 
 object FiberInterruptSpec extends ZIOBaseSpec {