--- a/core-tests/shared/src/test/scala/zio/internal/FiberInterruptSpec.scala
+++ b/core-tests/shared/src/test/scala/zio/internal/FiberInterruptSpec.scala
@@ -1,16 +1,16 @@
 package zio.internal
 
 import zio.Promise.internal.Pending
-import zio.clock.Clock
-import zio.test.Assertion.equalTo
-import zio.test.{ assert, suite, testM }
-import zio.{ Promise, ZIOBaseSpec }
 import zio.duration._
+import zio.test.Assertion.equalTo
+import zio.test._
+import zio.test.environment.Live
+import zio.{ Promise, ZIOBaseSpec }
 
 object FiberInterruptSpec extends ZIOBaseSpec {
 
   //Flaky
-  override def spec =
+  override def spec: ZSpec[Environment, Failure] =
     suite("FiberInterruptSpec")(
       testM("must interrupt all Promise joiners") {
 
@@ -18,7 +18,7 @@ object FiberInterruptSpec extends ZIOBaseSpec {
          * The idea here to spawn and interrupt as much joiners as possible
          * And to make sure that, all of them got cleaned up
          */
-        (for {
+        val io = for {
           p <- Promise.make[Throwable, Unit]
           loopFiber <- (for {
                          fb <- p.await.fork
@@ -30,10 +30,9 @@ object FiberInterruptSpec extends ZIOBaseSpec {
                           case _                => 0
                         }.delay(1.second) //Just to make sure all races already resolved
           _ <- p.succeed(())
+        } yield joinerSize
 
-        } yield {
-          assert(joinerSize)(equalTo(0))
-        }).provideCustomLayer(Clock.live)
+        assertM(Live.live(io))(equalTo(0))
       }
     )
 }