--- a/core/shared/src/main/scala/zio/internal/FiberContext.scala
+++ b/core/shared/src/main/scala/zio/internal/FiberContext.scala
@@ -16,7 +16,7 @@
 
 package zio.internal
 
-import java.util.concurrent.atomic.{ AtomicBoolean, AtomicReference }
+import java.util.concurrent.atomic.{ AtomicBoolean, AtomicLong, AtomicReference }
 
 import scala.annotation.{ switch, tailrec }
 import scala.collection.JavaConverters._
@@ -64,13 +64,13 @@ private[zio] final class FiberContext[E, A](
   private[this] val traceEffects: Boolean =
     traceExec && platform.tracing.tracingConfig.traceEffectOpsInExecution
 
-  private[this] val stack                 = Stack[Any => IO[Any, Any]]()
-  private[this] val environments          = Stack[AnyRef](startEnv)
-  private[this] val executors             = Stack[Executor](startExec)
-  private[this] val interruptStatus       = StackBool(startIStatus.toBoolean)
-  private[this] val supervisors           = Stack[Supervisor[Any]](supervisor0)
-  private[this] val forkScopeOverride     = Stack[Option[ZScope[Exit[Any, Any]]]]()
-  private[this] val evaluationsInProgress = new AtomicReference[EvaluationState](EvaluationState.Empty)
+  private[this] val stack              = Stack[Any => IO[Any, Any]]()
+  private[this] val environments       = Stack[AnyRef](startEnv)
+  private[this] val executors          = Stack[Executor](startExec)
+  private[this] val interruptStatus    = StackBool(startIStatus.toBoolean)
+  private[this] val supervisors        = Stack[Supervisor[Any]](supervisor0)
+  private[this] val forkScopeOverride  = Stack[Option[ZScope[Exit[Any, Any]]]]()
+  private[this] val evaluationThreadId = new AtomicLong(Long.MinValue)
 
   var scopeKey: ZScope.Key = null
 
@@ -276,13 +276,15 @@ private[zio] final class FiberContext[E, A](
   def evaluateNow(io0: IO[E, Any]): Unit = {
 
     // We must respect the order of instructions if another evaluation is in progress of current FiberContext
-    // Since we are not allowed to interrupt in the middle of another evaluation due to various race conditions,
-    // we will delay new zio effect and wait until current instruction will be completed by `shouldInterrupt()`
-    val currentEvaluation = evaluationsInProgress.get()
-    val newEvaluation     = EvaluationState.Pending(Thread.currentThread().getId)
+    // Since we are not allowed to interrupt it from another thread in the middle
+    // of another evaluation due to various race conditions,
+    // we will delay new zio effect and wait until current instruction will be completed
+    val contextThreadId = evaluationThreadId.get()
+    val currentThreadId = Thread.currentThread().getId
     if (
-      !currentEvaluation.isEqual(newEvaluation) ||
-      !evaluationsInProgress.compareAndSet(currentEvaluation, newEvaluation)
+      (contextThreadId != Long.MinValue &&
+        contextThreadId != currentThreadId) ||
+      !evaluationThreadId.compareAndSet(contextThreadId, currentThreadId)
     ) {
       evaluateLater(io0)
     } else {
@@ -673,7 +675,7 @@ private[zio] final class FiberContext[E, A](
           }
         }
       } finally {
-        evaluationsInProgress.set(EvaluationState.Empty)
+        evaluationThreadId.set(Long.MinValue)
         Fiber._currentFiber.remove()
       }
 
@@ -1091,25 +1093,6 @@ private[zio] final class FiberContext[E, A](
 
 }
 private[zio] object FiberContext {
-
-  sealed trait EvaluationState {
-    def isEqual(that: EvaluationState): Boolean
-  }
-
-  object EvaluationState {
-
-    case class Pending(threadId: Long) extends EvaluationState {
-      override def isEqual(that: EvaluationState): Boolean = that match {
-        case Pending(thatThreadId) if thatThreadId != threadId => false
-        case _                                                 => true
-      }
-    }
-
-    case object Empty extends EvaluationState {
-      override def isEqual(that: EvaluationState): Boolean = true
-    }
-
-  }
 
   sealed abstract class FiberState[+E, +A] extends Serializable with Product {
     def interrupted: Cause[Nothing]