@jdegoes 2020-12-07
@jsfwa `evaluateNow` should never be called concurrently. The fact it is being called concurrently means we have a bug somewhere, which should be fixed directly.

I will have a chance to look at this in a week.