@aartigao 2021-06-08
Hey @jdegoes, we are seeing a similar pattern in one of our services that use ZIO Kafka. Did you have a chance to look into this?

I tried to understand @jsfwa code in this PR but I'm afraid it's too low level for me. But if you need some heap dumps or something that can help you, just ask, I'll be willing to help in anything that help to solve this issue 👍🏽