@jdegoes 2021-06-08
`evaluateNow` cannot be called concurrently (except as noted above), and the solution to any problem in that logic cannot be to synchronize on `evaluateNow` (that's just working around a bug in a way that negatively impacts performance).

I believe @adamgfraser did work to fix one bug in this area. We should investigate to see if that fixes this, and if not, then base work off a minimal reproducer that can demonstrate concurrent calls to `evaluateNow`.